# packages/contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tula Auth API",
    "description": "Self-hostable authentication and user management. Client routes (`/v1/client/*`) take a publishable key; admin routes (`/v1/admin/*`) take a secret key. Every error body is an `ErrorEnvelope` with a stable `code`.",
    "version": "0.0.0"
  },
...
  "components": {
    "securitySchemes": {
      "publishableKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-tula-publishable-key",
        "description": "Publishable key `tula_pk_<env>_…`. Safe to embed in apps."
      },
      "accessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "EdDSA access token from a Tula session."
      },
      "secretKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Secret key `tula_sk_<env>_…`. Server-side only."
      }
    },
    "schemas": {
...
      "ErrorParams": {
        "type": "object",
        "propertyNames": {
          "type": "string"
        },
        "additionalProperties": {
          "type": [
            "string",
            "number",
            "boolean"
          ]
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "type": "string"
          },
          "params": {
            "$ref": "#/components/schemas/ErrorParams"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        },
        "required": [
          "status",
          "code",
          "detail"
        ]
      },
...
      "SessionClient": {
        "type": "string",
        "enum": [
          "web",
          "ios",
          "android",
          "server"
        ]
      },
...
      "FlowKind": {
        "type": "string",
        "enum": [
          "sign_in",
          "sign_up",
          "password_reset"
        ]
      },
      "EmailVerificationStrategy": {
        "type": "string",
        "enum": [
          "email_code",
          "email_link"
        ]
      },
      "SecondFactorMethod": {
        "type": "string",
        "enum": [
          "totp",
          "passkey",
          "backup_code",
          "sms_code"
        ]
      },
      "FlowStep": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "const": "needs_identifier"
              }
            },
            "required": [
              "status"
            ]
          },
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "const": "needs_password"
              }
            },
            "required": [
              "status"
            ]
          },
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "const": "needs_email_verification"
              },
              "destination": {
                "type": "string"
              },
              "strategies": {
                "minItems": 1,
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EmailVerificationStrategy"
                }
              }
            },
            "required": [
              "status",
              "destination",
              "strategies"
            ]
          },
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "const": "needs_new_password"
              },
              "destination": {
                "type": "string"
              },
              "strategies": {
                "minItems": 1,
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EmailVerificationStrategy"
                }
              }
            },
            "required": [
              "status",
              "destination",
              "strategies"
            ]
          },
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "const": "needs_second_factor"
              },
              "options": {
                "minItems": 1,
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SecondFactorMethod"
                }
              }
            },
            "required": [
              "status",
              "options"
            ]
          },
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "const": "complete"
              },
              "userId": {
                "type": "string"
              },
              "sessionId": {
                "type": "string"
              }
            },
            "required": [
              "status",
              "userId",
              "sessionId"
            ]
          }
        ]
      },
      "FlowAttempt": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/FlowKind"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "step": {
            "$ref": "#/components/schemas/FlowStep"
          },
          "session": {
            "$ref": "#/components/schemas/SessionTokens"
          }
        },
        "required": [
          "id",
          "kind",
          "expiresAt",
          "step"
        ]
      },
      "SignUpRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 320
          },
          "password": {
            "type": "string",
            "maxLength": 1024
          },
          "firstName": {
            "type": "string",
            "maxLength": 100
          },
          "lastName": {
            "type": "string",
            "maxLength": 100
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "SignInStartRequest": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "maxLength": 320
          }
        },
        "required": [
          "identifier"
        ]
      },
      "PasswordAttemptRequest": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "maxLength": 1024
          }
        },
        "required": [
          "password"
        ]
      },
...
      "PasswordResetStartRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 320
          }
        },
        "required": [
          "email"
        ]
      },
...
    }
  },
  "paths": {
...
    "/v1/client/sign-ups": {
      "post": {
        "operationId": "startSignUp",
        "tags": [
          "Flows"
        ],
        "summary": "Start a sign-up",
        "description": "Checks the email and password and emails a 6-digit code. The account is created when the code is verified. The response is the same whether or not the address already has an account. Send `x-tula-client` (`web`, `ios`, `android` or `server`) to choose how tokens are delivered when the flow completes.",
        "security": [
          {
            "publishableKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "The attempt, waiting on email verification.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowAttempt"
                }
              }
            }
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "array",
                      "items": {}
                    },
                    "data": {}
                  },
                  "required": [
                    "success",
                    "error",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "The request body is larger than 64 KiB (`request.too_large`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "The input is invalid; `errors` lists each field.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited; retry after `params.retryAfter` seconds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "A dependency is unreachable and the request was refused (`service.unavailable`); retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "x-tula-client",
            "schema": {
              "$ref": "#/components/schemas/SessionClient"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpRequest"
              }
            }
          }
        }
      }
    },
    "/v1/client/sign-ins": {
      "post": {
        "operationId": "startSignIn",
        "tags": [
          "Flows"
        ],
        "summary": "Start a sign-in",
        "description": "Always answers `needs_password`, whether or not the identifier belongs to an account. Send `x-tula-client` to choose how tokens are delivered when the flow completes.",
        "security": [
          {
            "publishableKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "The attempt, waiting on a password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowAttempt"
                }
              }
            }
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "array",
                      "items": {}
                    },
                    "data": {}
                  },
                  "required": [
                    "success",
                    "error",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "The request body is larger than 64 KiB (`request.too_large`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "The input is invalid; `errors` lists each field.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited; retry after `params.retryAfter` seconds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "A dependency is unreachable and the request was refused (`service.unavailable`); retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "x-tula-client",
            "schema": {
              "$ref": "#/components/schemas/SessionClient"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInStartRequest"
              }
            }
          }
        }
      }
    },
    "/v1/client/sign-ins/{attemptId}/password": {
      "post": {
        "operationId": "submitSignInPassword",
        "tags": [
          "Flows"
        ],
        "summary": "Submit the password",
        "description": "Completes the sign-in, or moves it to `needs_email_verification` when the user’s email is not verified yet. Every failure is the same `auth.invalid_credentials`. When the step is `complete`, `session` carries the tokens: browsers (`x-tula-client: web`, the default) receive the refresh token as an httpOnly cookie, other clients in the body.",
        "security": [
          {
            "publishableKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "The next step.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowAttempt"
                }
              }
            }
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "array",
                      "items": {}
                    },
                    "data": {}
                  },
                  "required": [
                    "success",
                    "error",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "The resource does not exist in this environment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The change conflicts with existing data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "The request body is larger than 64 KiB (`request.too_large`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "The input is invalid; `errors` lists each field.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited; retry after `params.retryAfter` seconds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "A dependency is unreachable and the request was refused (`service.unavailable`); retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "attemptId",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordAttemptRequest"
              }
            }
          }
        }
      }
    },
...
    "/v1/client/password-resets": {
      "post": {
        "operationId": "startPasswordReset",
        "tags": [
          "Flows"
        ],
        "summary": "Start a password reset",
        "description": "Emails a 6-digit code to the address. The response is the same whether or not the address has an account. Send `x-tula-client` to choose how tokens are delivered when the reset completes.",
        "security": [
          {
            "publishableKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "The attempt, waiting on the code and a new password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowAttempt"
                }
              }
            }
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "array",
                      "items": {}
                    },
                    "data": {}
                  },
                  "required": [
                    "success",
                    "error",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "The request body is larger than 64 KiB (`request.too_large`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "The input is invalid; `errors` lists each field.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited; retry after `params.retryAfter` seconds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "A dependency is unreachable and the request was refused (`service.unavailable`); retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "x-tula-client",
            "schema": {
              "$ref": "#/components/schemas/SessionClient"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetStartRequest"
              }
            }
          }
        }
      }
    },
...
  }
}